refactor(utils): move getCropLifecycleDuration to its own file
