+++ src/components/Item/Item.js
@@ -19,13 +19,13 @@ import { moneyString } from '../../utils/moneyString'
 import {
   inventorySpaceRemaining,
   isItemSoldInShop,
-  getCropLifecycleDuration,
   getFinalCropItemFromSeedItem,
   getItemCurrentValue,
   getResaleValue,
   getSalePriceMultiplier,
   integerString,
 } from '../../utils'
+import {getCropLifecycleDuration} from "../../utils/getCropLifecycleDuration"
 import QuantityInput from '../QuantityInput'
 import AnimatedNumber from '../AnimatedNumber'
