+++ src/utils/getCropLifecycleDuration.js
@@ -1,0 +1,9 @@
+import { memoize } from './memoize'
+
+/**
+ * @param {farmhand.crop} crop
+ * @returns {number}
+ */
+export const getCropLifecycleDuration = memoize(({ cropTimetable }) =>
+  Object.values(cropTimetable).reduce((acc, value) => acc + value, 0)
+)
